+++ Config/Schema/schema.php
@@ -8,33 +8,7 @@ class AppSchema extends CakeSchema {
   }
 
   public function after($event = array()) {
-    if(isset($event['create'])) {
-      $sql_cmd = "";
-      switch($event['create']) {
-        case 'cm_link_org_identity_enrollers':
-          $LinkOrgIdentityEnroller = ClassRegistry::init('LinkOrgIdentityEnroller.LinkOrgIdentityEnroller');
-          $LinkOrgIdentityEnroller->useDbConfig = $this->connection;
-          $sql_cmd = "ALTER TABLE ONLY public.cm_link_org_identity_enrollers ADD CONSTRAINT cm_link_org_identity_enrollers_co_id_fkey FOREIGN KEY (co_id) REFERENCES public.cm_cos(id)";
-          // Add the constraints or any other initializations
-          $LinkOrgIdentityEnroller->query($sql_cmd);
-          break;
-        case 'cm_link_org_identity_eofs':
-          $LinkOrgIdentityEof = ClassRegistry::init('LinkOrgIdentityEnroller.LinkOrgIdentityEof');
-          $LinkOrgIdentityEof->useDbConfig = $this->connection;
-          $sql_cmd = "ALTER TABLE ONLY public.cm_link_org_identity_eofs ADD CONSTRAINT cm_link_org_identity_eofs_link_org_identity_enroller_id_fkey FOREIGN KEY (link_org_identity_enroller_id) REFERENCES public.cm_link_org_identity_enrollers(id);";
-          $sql_cmd .= "ALTER TABLE ONLY public.cm_link_org_identity_eofs ADD CONSTRAINT cm_link_org_identity_eofs_co_enrollment_flow_id_fkey FOREIGN KEY (co_enrollment_flow_id) REFERENCES public.cm_co_enrollment_flows(id);";
-          // Add the constraints or any other initializations
-          $LinkOrgIdentityEof->query($sql_cmd);
-          break;
-        case 'cm_link_org_identity_states':
-          $LinkOrgIdentityState = ClassRegistry::init('LinkOrgIdentityEnroller.LinkOrgIdentityState');
-          $LinkOrgIdentityState->useDbConfig = $this->connection;
-          $sql_cmd = "ALTER TABLE ONLY public.cm_link_org_identity_states ADD CONSTRAINT cm_link_org_identity_states_link_org_identity_enroller_id_fkey FOREIGN KEY (link_org_identity_enroller_id) REFERENCES public.cm_link_org_identity_enrollers(id);";
-          // Add the constraints or any other initializations
-          $LinkOrgIdentityState->query($sql_cmd);
-          break;
-      }
-    }
+    return true;
   }
 
   public $link_org_identity_enrollers = array(
@@ -88,12 +62,9 @@ class AppSchema extends CakeSchema {
 
 
 /**
- * // Console/cake schema create --file schema.php --path /srv/comanage/registry-current/local/Plugin/LinkOrgIdentityEnroller/Config/Schema
- *
- * ALTER TABLE ONLY public.cm_link_org_identity_enrollers ADD CONSTRAINT cm_link_org_identity_enrollers_co_id_fkey FOREIGN KEY (co_id) REFERENCES public.cm_cos(id);
- *
- * ALTER TABLE ONLY public.cm_link_org_identity_states ADD CONSTRAINT cm_link_org_identity_states_link_org_identity_enroller_id_fkey FOREIGN KEY (link_org_identity_enroller_id) REFERENCES public.cm_link_org_identity_enrollers(id);
- *
- * ALTER TABLE ONLY public.cm_link_org_identity_eofs ADD CONSTRAINT cm_link_org_identity_eofs_link_org_identity_enroller_id_fkey FOREIGN KEY (link_org_identity_enroller_id) REFERENCES public.cm_link_org_identity_enrollers(id);
- * ALTER TABLE ONLY public.cm_link_org_identity_eofs ADD CONSTRAINT cm_link_org_identity_eofs_co_enrollment_flow_id_fkey FOREIGN KEY (co_enrollment_flow_id) REFERENCES public.cm_co_enrollment_flows(id);
+ // Console/cake schema create --file schema.php --path /srv/comanage/registry-current/local/Plugin/LinkOrgIdentityEnroller/Config/Schema
+ ALTER TABLE ONLY cm_link_org_identity_enrollers ADD CONSTRAINT cm_link_org_identity_enrollers_co_id_fkey FOREIGN KEY (co_id) REFERENCES cm_cos(id);
+ ALTER TABLE ONLY cm_link_org_identity_states ADD CONSTRAINT cm_link_org_identity_states_link_org_identity_enroller_id_fkey FOREIGN KEY (link_org_identity_enroller_id) REFERENCES cm_link_org_identity_enrollers(id);
+ ALTER TABLE ONLY cm_link_org_identity_eofs ADD CONSTRAINT cm_link_org_identity_eofs_link_org_identity_enroller_id_fkey FOREIGN KEY (link_org_identity_enroller_id) REFERENCES cm_link_org_identity_enrollers(id);
+ ALTER TABLE ONLY cm_link_org_identity_eofs ADD CONSTRAINT cm_link_org_identity_eofs_co_enrollment_flow_id_fkey FOREIGN KEY (co_enrollment_flow_id) REFERENCES cm_co_enrollment_flows(id);
  */