+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [v.0.3.1] - 2020-11-18
+### Added
+- Configuration shell script for setup
+- SQL script containing tables and constraints required for the plugin
+
 ## [v0.3.0] - 2020-10-21
 ### Added
 - Added type of Linking(Implicit|Explicit) in LinkOrgIdentityState model. It will be used for keeping history records.